* big file overhaul again. Created main.py, moved rest into src/utils module. pytest and main.py both run successfully
