--- src/__init__.py
@@ -1,8 +1,0 @@
-from . import constants
-
-constants.IS_TEST = True
-
-from . import analysis
-from . import interfaces
-from . import routines
-from . import titration
