+++ src/utils/devices/temperature_probe.py
