+++ test/serial_mock_test.py
@@ -1,5 +1,5 @@
-import src.constants as c
-import src.devices.serial_mock as serial
+import src.utils.constants as c
+import src.utils.devices.serial_mock as serial
 
 
 def test_serial_create():
